% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{aggregate_trajectories}
\alias{aggregate_trajectories}
\title{Function to aggregate trajectories}
\usage{
aggregate_trajectories(
  traj_object,
  aggregate_days = "14 days",
  seamount_groups = "all"
)
}
\arguments{
\item{traj_object}{This is a data frame containing the output of daily drift trajectories}

\item{aggregate_days}{Time period over which to aggregate options are; 7 days, 14 days (default) and 1 month}

\item{seamount_groups}{Seamounts over which to group (default is all), individual seamounts can be used as well}
}
\description{
This function takes a set of daily larval drift positions and aggregates them by weeks or months of release. For example,
the function will average the daily latitude and longitude for increments of two week period of releases across all seamounts.
}
\examples{
test1<-aggregate_trajectories(alltraj,aggregate_days="1 month",seamount_groups="Koko")



}
\keyword{Larval}
\keyword{drift}
