% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AVHRR_daily_extract_function}
\alias{AVHRR_daily_extract_function}
\title{Function to extract the sea surface temperature at a date and location}
\usage{
AVHRR_daily_extract_function(
  Longitude,
  Latitude,
  date,
  destination_path,
  do_download = TRUE,
  Area_of_interest = c(-230, -120, 15, 60)
)
}
\arguments{
\item{Longitude}{Longitude of interest}

\item{Latitude}{Latitude of interest}

\item{date}{date of interest}

\item{destination_path}{Local path for location of AVHRR daily files}

\item{do_download}{TRUE (default) or FALSE should the AVHRR daily files be downloaded}

\item{Area_of_interest}{Either a polygon or x and y min and max defining the area of interest (defauilt is the North Pacific)}
}
\description{
This function downloads daily Optimum Interpolation Sea Surface Temperatures from the NCDC website (https://www.ncdc.noaa.gov/oisst)
and stores them to a local destination. It will then extract the sea surface temperature for a set of dates at a set of locations.
THe SST is gridded on 1/4 degrees, so the output will be on the same grid size.
}
\examples{
temps<-AVHRR_daily_extract_function(alltraj$Lon[1:4],alltraj$Lat[1:4],alltraj$date[1:4],"D:/AVHRR Data/",do_download=FALSE)



}
\keyword{AVHRR}
\keyword{Larval}
\keyword{drift,}
\keyword{sea}
\keyword{surface}
\keyword{temperature,}
