% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dist_xy}
\alias{dist_xy}
\title{Function to calculate distance}
\usage{
dist_xy(Longitude_start, Latitude_start, Longitude_end, Latitude_end, unit)
}
\arguments{
\item{Longitude_start}{Start longitude}

\item{Latitude_start}{Start latitude}

\item{Longitude_end}{End longitude}

\item{Latitude_end}{End latitude}

\item{unit}{units to use for calculation (either m or km)}
}
\description{
This function calculates the distance between two points on a round earth.
}
\examples{
dist_xy(-120,58,-121,59,"km")
}
\keyword{calculation}
\keyword{distance}
