% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MODIS_extract_function}
\alias{MODIS_extract_function}
\title{Function to extract the Modis data at a date and location}
\usage{
MODIS_extract_function(
  Longitude,
  Latitude,
  date,
  destination_path,
  do_download = TRUE,
  Area_of_interest = c(-230, -120, 15, 60)
)
}
\arguments{
\item{Longitude}{Longitude of interest}

\item{Latitude}{Latitude of interest}

\item{date}{date of interest}

\item{destination_path}{Local path for location of MODIS daily files}

\item{do_download}{TRUE (default) or FALSE should the MODIS daily files be downloaded}

\item{Area_of_interest}{Either a polygon or x and y min and max defining the area of interest (defauilt is the North Pacific)}
}
\description{
This function downloads MODIS satellite ocean productivity from the Oregon State University Ocean Productivity website
(http://sites.science.oregonstate.edu/ocean.productivity/standard.product.php) and stores them to a local destination.
It will then extract the chlorophyll data for a set of dates at a set of locations and merge to the original data set.
The satellite data is only available on 8 day intervals, so the output will contain NA where no data were available.
THe Ocean Color is gridded on 1/4 degrees in this output.
}
\examples{
temps<-MODIS_daily_extract_function(alltraj$Lon[1:4],alltraj$Lat[1:4],alltraj$date[1:4],"D:/MODIS data/",do_download=FALSE)



}
\keyword{AVHRR}
\keyword{Larval}
\keyword{drift,}
\keyword{sea}
\keyword{surface}
\keyword{temperature,}
