% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Particle_drift}
\alias{Particle_drift}
\title{Function to extract start-end positions and distance}
\usage{
Particle_drift(nc_file, drift_days)
}
\arguments{
\item{nc_file}{Output ncdf file from Ocean Parcels}

\item{drift_days}{number of days for which to calculate the drift}
}
\description{
This function reads a net cdf file output from Ocean Parcels and compiles the final position of the particle, as well as the start position,
beginning and ending dates, the cumulative distance travelled and the straight-line distance travelled. It also attaches the year of release
and the seamount of release to the output.
}
\examples{
Particle_drift(kkjkh, 120)
}
\keyword{drift}
\keyword{duration}
\keyword{globcurrent}
\keyword{larval}
