% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Particle_trajectory}
\alias{Particle_trajectory}
\title{Function to extract start-end positions and distance}
\usage{
Particle_trajectory(nc_file, drift_days, interval = 1)
}
\arguments{
\item{nc_file}{Output ncdf file from Ocean Parcels}

\item{drift_days}{number of days for which to calculate the drift}

\item{interval}{interval in number of days over which to output the larval position (e.g. daily (the default), weekly = 7)}
}
\description{
This function reads a net cdf file output from Ocean Parcels and outputs the trajectory of the particals on an interval in days. The function
also attaches the year of release and the seamount of release to the output.
}
\examples{
Particle_trajectory(kkjkh, 120, 7)
}
\keyword{drift}
\keyword{duration}
\keyword{globcurrent}
\keyword{larval}
